Frontend tf.math fixes (#4928)
* remove code duplication

* Fix Instance methods

* fixes to tf.math

Former-commit-id: 0f5b39e9fca675a5e67d475897a98f2d9ccfca02

+++ ivy/functional/frontends/tensorflow/math.py
@@ -157,7 +157,7 @@ def polyval(coeffs, x, name=None):
     ivy.assertions.check_isinstance(coeffs, list)
     x = ivy.array(x)
     if len(coeffs) < 1:
-        return ivy.zeros_like(x)
+        return ivy.zeros_like(x, dtype=x.dtype)
     coeffs = [ivy.array(_) for _ in coeffs]
     p = coeffs[0]
     for c in coeffs[1:]:
